add links to onboarding in Graphbook, add step to set up 1-1s (sourcegraph/about#326)
I wanted to add a step to the onboarding checklist for all new teammates (not just engineers) to set up 1-1s with their manager. I saw that this was mentioned in the Graphbook. Instead of submitting a PR to the Graphbook (which will soon be migrated to this handbook) to add this, I added it here and added some links back to the Graphbook.