define IARR, new/expansion IARR (sourcegraph/about#546)
* define IARR, new/expansion IARR

I propose we use IARR (defined in this change) instead of "net new ARR" as we have been saying. This is to avoid confusion; "new ARR" can be interpreted as "incremental ARR" or "new logo ARR", which are 2 very different things. Now that we are distinguishing between new and expansion ARR much more than in the past, this confusion is more likely to occur.

Note (for the curious): GitLab uses the terms "new IACV" and "growth IACV" (https://about.gitlab.com/handbook/sales/#annual-contract-value-acv). We use the term "expansion" instead of "growth". If anyone has a preference for "growth" instead, speak up.

* net new ARR -> IARR