CEO no longer interviews every candidate (sourcegraph/about#2344)
I removed myself from the interview process for some (but not all) roles. The reason for this is that I simply don't have the time to interview everyone anymore. To decide which roles to remove myself from, I chose the roles where I would work less directly with the person and where we have successfully hired and onboarded more people in that role already (which indicates to me that my input is less needed).