New process for customer names in public places (avoids needing hubspot/salesforce access) (sourcegraph/about#3469)
We previously used hubspot and/or salesforce links as just a private customer names mapping index, but as we grow it doesn't make sense to pay for company-wide access to these tools just for this purpose. To come up with an inclusive and much cheaper alternative, we've aligned on using a private google sheet (as you can link to individual cells).