Update engineering organization
As of 2021-09-13 we reorganized the product and engineering teams (internal document) so some team pages are missing and need to be created. Until the handbook is updated, the source of truth for team assignments and hiring plans are in this spreadsheet (internal document).

This removes the autogenerated chart because:

There is no way to make that accurate without me changing all the team pages (I want to distribute this work)
It was annoying that it was frequently slow to load
It was brittle because anchors it depended on kept getting removed
Current SoT is the spreadsheet, so want to delete inaccurate info.
We can decide to add it back after the handbook is updated if we want, or we can consider other solutions moving forward.