Adding step for Engineering Hiring Managers (#536)
* Adding step for Engineering Hiring Managers

Adding an additional step for Hiring Managers that creates greater visibility into final hiring decisions with the broader interviewing team. This will help calibrate for future interviews. 

I think this could make sense as the last item in Step 2 or could live in the Interview Debriefs section below. Open to feedback.

* squash! Prettier

* Update content/talent/resources_for_hiring_managers.md

Co-authored-by: Nick Snyder <nick@sourcegraph.com>

* Update resources_for_hiring_managers.md

* Update resources_for_hiring_managers.md

Per Nick's feedback and Devon's approval, removed the "Interview debriefs" section since it is covered in the new step we added.

* squash! Prettier

Co-authored-by: msennett22 <msennett22@users.noreply.github.com>
Co-authored-by: Nick Snyder <nick@sourcegraph.com>