delete job description template & guidelines page (#663)
The template listed here is linked to from this page: https://github.com/sourcegraph/handbook/blob/main/content/talent/hiring/index.md

I'm proposing deleting this page because it contains duplicated information that exists elsewhere. Additionally, this page is not linked to from anywhere in the Handbook today. This is preventing us from re-implementing a Handbook check requiring all pages to be linked to within the Handbook. If we would like to keep this page, we need to link to it from somewhere in the Handbook.