Removed to TBDs to make this easier to read
Dan keep or toss my recommendations here.  I went ahead and cut out all the content that was TO DO or TBD  just to make the page easier to read.  As we flush out more content we can add it back in.  My intent here was to make keep complete info out there.