Sourcegraphers use case for Batch Changes (#1141)
From time to time, we hear about Sourcegrapher/Dogfood use cases in the #batch-changes channel. There's no great source of truth for storing this, but I think this is valuable data for us and for potential teammates that are considering using Batch Changes.
I don't know where this should live, added here by default but open to suggestions.