Add defensibility and tech investment
I think we are missing a section explaining how we invest in immediate value features, but also have a long-term strategy to build foundational technology to explore and manipulate the Sourcegraph.
In practice: the day Code Intel and Compute are available as a platform to, say, create batch changes, we will be able to unlock a wow! use case AND this will be hard to copy because it is a lot of foundational work and a lot of longing operational work to provide this layer.