Separate problem and solution review in RFCs
A pattern I've noticed is that some RFCs jump into detailed solutions prior to the problem that's being solved is well understood; sometimes the problem doesn't exist, sometimes the problem is just a little bit different. This separates solutioning from validating the problem so that we can be more efficient and avoid getting ahead of ourselves too much.

In situations where the solution is written too early, there's often a ton of debate on the details of the solution that would have been more managable if it was around the problem.