core: precise wording of monorepo performance goal (#2778)
We now have a repository which fits the bill of our older performance goal: gigarepo. Lets directly reference it to make the goal easier to communicate. I include what the gigarepo is for people not familiar with the repository. Additionally, I found the wording "15GB repo" unclear since there are many ways to interpret that. So this specifically states it is the HEAD working copy size. IE it excludes the .git folder size.

Maybe we can explain what gigarepo is somewhere else rather than in the bullet points? For now this makes it more self contained.

Co-authored-by: Eugenia Fernandez <eugeniaft@users.noreply.github.com>