Update bug bounty scheme
Remove the previous scheme description. We can't point to the current
scheme because it's private, so explain that people who submit good
vulnerabilities will be considered for an invite, but that any further
use of the email address doesn't constitute a submission to the program.

I think this should cover all bases!