update team and org chart (#3789)
After the product/engineering reorg (eg Cloud -> Admin Experience), I did my best to make a big update that moves things into the right directories, etc. If anything is wrong here from what was communicated, the mistake is in this commit and not in what was communicated. Please submit a PR to fix it.