Add a security incident response process (#4106)
* Add a security incident response process

This complements the security incident response policy, defining some clear
steps on exactly how we should be handling incidents in the future.

I've chosen to use the private Google Drive folder instead of incident.io as we
have no way of enabling private incident.io incidents due to our Slack setup. It's
not wise to have security incidents public by default, so we should work on them
privately until we have deemed them safe to share. I have noted – in line with
our values on transparency – that we should share these documents within the organiztion
if we deem it safe, once an incident has been resolved.

* squash! Prettier

Co-authored-by: ferozsalam <ferozsalam@users.noreply.github.com>