Proposal: remove IC5 profiles
This proposal is the result of yesterday's and last week's IC5 sync in
which we agreed that the IC5 profiles cause us more confusion than
anything else. Confusion comes from us asking us the following questions
constantly:

- are the profiles a set of expectations that we need to live up to?
- what if we want to do something different, something that doesn't
  match the profile anymore? does the profile need to be rewritten? do
  we lose the title of IC5?
- in some cases the profiles were reverse-engineering descriptions of
  what individuals were already doing - does that make the profiles
  unfit to lead and serve as a north star?
- in other cases the profiles are really abstract and neither serve as a
  north star nor as a description that other ICs can look to as guidance
  for what IC5-level work looks like
- ...

We do not yet know how to further shape the IC5 role, whether we'll find
something to replace the IC5 profiles or not, whether we'll adopt
industry-standard terminology (staff eng & archetypes such as: tech
lead, architect, ...), or do something else.

This is meant as a first step _back_ so we can decide which two steps
forward we want to take.

It's a *proposal* too: this PR should serve as a discussion and if
someone disagrees with this proposal and we end up having a discussion
around IC5 profiles, it served its purpose.

If we (the IC5s) all agree on the removal, we'll ask Quinn for his approval.