Update research compensation process (#4577)
* Update research compensation process

Following a conversation with @desenesterling from the Finance team, we have updated the research compensation process to align with Finance team processes.

* squash! Prettier

* Update handbook URL link

* Updating the handbook URL again

Sorry I am bad at GitHub. Note: the suggested URL in the error message was wrong. I had to edit it myself to delete the last slash so that .md was not disrupted.

Co-authored-by: meglynst <meglynst@users.noreply.github.com>
Co-authored-by: Steve Fisher <steve.fisher@sourcegraph.com>