document engineering feature considerations
These are the things we should consider for everything we ship (new and existing features).

I've had this list in my mind for a couple weeks and wanted to write it down so that we can (as an eng dept) align on these considerations/principles.

Eventually this could become something like a checklist, but for now (until we've worked with these a bit more), let's keep it as high-level principles.