Lift Slack retention default policy to 180 days (#5731)
This is a temporary change to alleviate some pressure felt by the business. We may eventually lift the short policy entirely, we may tighten it again—but we need to define answers to how we will communicate as a team going forward first. 

So yes, this is kicking the can down the road by 3 months :)