equality of opportunity
We work hard to recruit, reward, and support the best people from all over the world to be part of the Sourcegraph team, including people who have varying degrees of access to opportunities. We hire and reward people based on their contributions to our mission and our company. We value, respect, and support all team members, and we celebrate that our team has all kinds of backgrounds, personal characteristics, beliefs, ideologies, and personal values. All of this is critical for us to make it so everyone codes (our mission).