cloud/v2: best-effort on-prem-to-cloud data migration refresh (#6980)
A simple best-effort refresh to the old on-prem data migrations guide
for Cloud v2 that I've been promising for a while. Where an equivalent
in Cloud v2 exists, I've updated the guide for the equivalent - where
one doesn't, I've left a `TODO` indicating as such. Pending further
customer interest/requirement for a data migration there are no plans at
the moment to fill in the gaps.

This process will be owned by @sourcegraph/implementation-eng but Cloud
is still committed to supporting the first few migrations if they ever
come up, and since the process pertains considerably to Cloud tooling
and infrastructure I think the guide should live in the `/cloud`
handbook pages.

Part of https://github.com/sourcegraph/customer/issues/1525