First stab at updates for Source, APIs, Search Platform
This updates the `.yml` files so that all the relationships and roles
are up-to-date.

It also updates the team pages - at least somewhat. The "Search Core"
team page still needs to be moved/renamed, but I didn't want to change
too many links.

Still need to do more revisions, but I don't want to add too much into a
single PR.