Add "Successful engineers at Sourcegraph" document
I added the document as-is in Markdown form and then linked to it from
the index.

I also think it should be linked-to on the ladder page (because that's
what most people look at), but I think that page will be reworked anyway
(should be, I think), so I only added 2 clumsy paragraphs.