Chore: consolidate code search teams (#8249)
This consolidates the teams that are now under the Code Search umbrella
and removes a bunch of outdated content. This is a re-implementation of
https://github.com/sourcegraph/handbook/pull/8083, but without the name
change to Search Suite.

There is definitely still more to do to clean up here, but I figure it's
a good start and will make the handbook a little less confusing for
people looking for information.