Update invoices.md to account for currency conversion fees
I faced a roughly 1.5% in conversion fee because I thought that I would be paid the salary in my currency which is Canadian Dollar. 

I got that impression from the existing description in this page. So now I am editing the page to help other people to avoid facing this.